Fix issue with <select> nodes
Whenever options get removed, other options gets selected.

This is problematic in some case, but luckily that does not
happen when the select insertBefore option nodes.

This change reduce a bit overall code size, and it might
improve performance in most common cases/browsers.

It also fixes a nasty issue in lighterhtml:
https://github.com/WebReflection/lighterhtml/issues/50